Checking functionality of playlist view routing (using navigation links for these sub-views)
